Set up Creep Base, Mouse Cursor (Visbility) and new architecture for towers
Initially had a lot of ideas on how I should classify tower/projectiles so it has maximum versatility and we could use this system to create anything we want without touching the code much. But then I decided to prebuild something's into tower base rather in inherit as some functions will be commonly used and rather than inherit a new field, why not just create presets on the base and allow users to select the options.