Tower Radius Creep Detection
After lots of experimenting and R&D I have created the detection system within the Tower Radius. Interestingly, if we build a tower with the creep in range, it wouldn't be able to register it in its range. Hence, I added another functionality at the start to make sure that it does a pre-scan once and aqquires all target. Everything after that is then registered and deregistered with the detection system.