Behaviour System for Projectile Effects
To plan for a more fluid code architecture, this new system will now be put into place. In the future, using this will make it alot easier for us to design new towers/add new features/towers whatever. We just need to attach a component which derives from UProjectileEffectBase which allows us to mark the effect to trigger on update or on hit. With these two, we can craft almost anything without having to rewrite code to fit in new requirements.

Examples
-> Creating a tower with expanding shots
Add a UProjectileEffectBase derived script which overrides the function ApplyEffect and make it scale actor size up. Change projectileEffectType to update.

-> Create a tower which performs AOE on hit
Add a UProjectileEffectBase derived script which overrides the function ApplyEffect and make it create a SphereCast of certain radius (Which we could make it a custom value with the script). Change projectileEffectType to onhit.